fix(declare): refine varname validation
